+++ docs/changelog.md
@@ -14,13 +14,16 @@ title: Changelog
 
 ## [0.1.2] - unreleased
 ### Added
-- improve demo site
 - add debug logging handler
-- improved component import error handling
 - allow component names to be dynamic
 - @v templatetag for "live" rendering of frontend variables
 - support Django models in TetraJSONEn/Decoder
 - experimental FormComponent and GenericObjectFormComponent support with form validation
+
+### Changed
+- more verbose error when template is not enclosed in HTML tags
+- improved component import error handling
+- improve demo site
 
 ## [0.1.1] - 2024-04-10
 ### Changes